# Conflicts:
#	src/main/java/sgpiv/config/NotificacionInterceptor.java
#	src/main/java/sgpiv/config/WebConfig.java
#	src/main/java/sgpiv/controller/NotificacionController.java
#	src/main/java/sgpiv/model/Notificacion.java
#	src/main/java/sgpiv/repository/NotificacionRepository.java
#	src/main/java/sgpiv/service/NotificacionService.java
#	src/main/java/sgpiv/service/SolicitudService.java
#	src/main/resources/application.propierties.example
#	src/main/resources/templates/fragments/sidebar.html
#	src/main/resources/templates/notificaciones.html

+++ src/main/java/sgpiv/dtos/request/LoteRequestDTO.java
@@ -20,7 +20,7 @@ public class LoteRequestDTO {
 
     @NotNull(message = "Ingrese la superficie")
     @Positive(message = "La superficie debe ser mayor a 0")
-    private Float superficie;
+    private Double superficie;
 
     @NotBlank(message = "Ingrese la ubicacion")
     private String ubicacion;
@@ -35,8 +35,5 @@ public class LoteRequestDTO {
 
     private String restricciones;
 
-    private Set<ServicioLote> servicios =
-            new HashSet<>();
-
-
+    private Set<ServicioLote> servicios = new HashSet<>();
 }
